+++ cypress/integration/alert-spec.js
@@ -17,7 +17,7 @@ describe('Stateless alert', () => {
     cy.contains('a', 'Say Hi')
   })
 
-  it.skip('alerts with name', () => {
+  it('alerts with name', () => {
     cy.contains('Say Hi').click()
     cy.get('@alert').should('have.been.calledOnce')
   })
