+++ cypress/integration/pure-component.spec.js
@@ -3,6 +3,7 @@
 
 import React from 'react'
 import Button from '../../src/pure-component.jsx'
+import { mount } from 'cypress-react-unit-tests'
 
 /* eslint-env mocha */
 describe('Button pure component', function () {
