+++ tsconfig.json
@@ -59,5 +59,5 @@
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
     "esModuleInterop": true
   },
-  "include": ["lib/index.ts", "lib/index.d.ts"]
+  "include": ["lib/**/*.ts"]
 }
